change the architecture from rv32 to rv64

+++ src/paging.rs
@@ -1,8 +1,8 @@
-pub static mut NEXT_PADDR: u32 = 0;
+pub static mut NEXT_PADDR: u64 = 0;
 
-pub unsafe fn alloc_pages(n: u32) -> crate::types::PaddrT {
+pub unsafe fn alloc_pages(n: u64) -> crate::types::PaddrT {
     let paddr = NEXT_PADDR;
-    NEXT_PADDR += n * crate::constants::PAGE_SIZE as u32;
+    NEXT_PADDR += n * crate::constants::PAGE_SIZE;
 
     if NEXT_PADDR > core::ptr::addr_of!(crate::__free_ram_end) as crate::types::PaddrT {
         panic!("out of memory");
@@ -11,7 +11,7 @@ pub unsafe fn alloc_pages(n: u32) -> crate::types::PaddrT {
     core::ptr::write_bytes(
         paddr as *mut u8,
         0,
-        (n * crate::constants::PAGE_SIZE as u32) as usize,
+        (n * crate::constants::PAGE_SIZE) as usize,
     );
 
     paddr
@@ -21,31 +21,28 @@ pub unsafe fn map_page(
     table1: crate::types::PaddrT,
     vaddr: crate::types::VaddrT,
     paddr: crate::types::PaddrT,
-    flags: u32,
+    flags: u64,
 ) {
-    if vaddr % crate::constants::PAGE_SIZE as u32 != 0 {
+    if vaddr % crate::constants::PAGE_SIZE != 0 {
         panic!("unaligned vaddr {:x}", vaddr);
     }
 
-    if vaddr % crate::constants::PAGE_SIZE as u32 != 0 {
+    if vaddr % crate::constants::PAGE_SIZE != 0 {
         panic!("unaligned paddr {:x}", paddr);
     }
 
-    let table1 = table1 as *mut u32;
+    let table1 = table1 as *mut u64;
     let vpn1 = ((vaddr >> 22) & 0x3ff) as isize;
     if *table1.offset(vpn1) & crate::constants::PAGE_V == 0 {
         let pt_paddr = alloc_pages(1);
-        table1.offset(vpn1).write(
-            (((pt_paddr / crate::constants::PAGE_SIZE as u32) as u32) << 10)
-                | crate::constants::PAGE_V,
-        );
+        table1
+            .offset(vpn1)
+            .write(((pt_paddr / crate::constants::PAGE_SIZE) << 10) | crate::constants::PAGE_V);
     }
 
-    let table0 = ((*table1.offset(vpn1) >> 10) * crate::constants::PAGE_SIZE as u32) as *mut u32;
+    let table0 = ((*table1.offset(vpn1) >> 10) * crate::constants::PAGE_SIZE) as *mut u64;
     let vpn0 = ((vaddr >> 12) & 0x3ff) as isize;
-    table0.offset(vpn0).write(
-        ((paddr / crate::constants::PAGE_SIZE as u32) << 10)
-            | flags
-            | crate::constants::PAGE_V,
-    );
+    table0
+        .offset(vpn0)
+        .write(((paddr / crate::constants::PAGE_SIZE) << 10) | flags | crate::constants::PAGE_V);
 }
